fix: move TaskRepository to infrastructure folder [sc-270]
