+++ server/src/application/use-cases/createTaskUseCase.ts
@@ -1,0 +1,15 @@
+import { Task, TaskProps } from '../../domain/entities/task'
+import { TaskRepository } from './../../infrastructure/persistence/TaskRepository'
+
+export class CreateTaskUseCase {
+  private taskRepository: TaskRepository
+
+  constructor(taskRepository: TaskRepository) {
+    this.taskRepository = taskRepository
+  }
+
+  async execute(taskData: TaskProps): Promise<Task> {
+    const task = this.taskRepository.create(taskData)
+    return await this.taskRepository.save(task)
+  }
+}
