+++ server/src/index.ts
@@ -1,0 +1,7 @@
+import { app } from './infrastructure/express/app';
+
+const PORT = process.env.PORT || 3000;
+
+app.listen(PORT, () => {
+  console.log(`Server is running on port ${PORT}`);
+})
