+++ server/src/infrastructure/express/app.ts
@@ -1,0 +1,11 @@
+import express from 'express';
+
+
+const app = express();
+app.use(express.json());
+
+const taskController = new TaskController();
+
+app.post('/tasks', (req, res) => taskController.createTask(req, res));
+
+export { app };
